refactor build schema
